backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
aiofiles==23.2.1

# Database
sqlalchemy==2.0.25
alembic==1.13.1
sqlite-utils==3.36

# AI/ML - Face Recognition
insightface==0.7.3
onnxruntime==1.16.3
onnxruntime-gpu==1.16.3

# AI/ML - CLIP and Vision
torch==2.1.2
torchvision==0.16.2
transformers==4.36.2
pillow==10.2.0
opencv-python==4.9.0.80

# Vector Search
faiss-cpu==1.7.4
# faiss-gpu==1.7.4  # Uncomment for GPU support

# Image Processing
imagehash==4.3.1
pillow-heif==0.15.0
rawpy==0.19.0

# Clustering
scikit-learn==1.4.0
hdbscan==0.8.33
numpy==1.26.3

# Utilities
python-dateutil==2.8.2
tqdm==4.66.1
pydantic==2.5.3
pydantic-settings==2.1.0

# Async & Background Tasks
celery==5.3.6
redis==5.0.1

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
flake8==7.0.0
httpx==0.26.0
